fix(create-wdio): emit driverProvider in tauri config, escape Windows binary path
- services.ejs: map TauriDriverProviderChoice enum to 'official'|'crabnebula'|
  'embedded' config string and emit it as driverProvider in the generated
  wdio.conf — previously the user's provider selection was used only for the
  post-install banner but never written into the config, so the service would
  silently default to 'official' regardless of choice
- services.ejs: use JSON.stringify() for tauriAppBinaryPath so Windows paths
  with backslashes are properly double-quoted and escaped in the generated file
  (raw single-quoted interpolation produces broken \t, \u etc. escape sequences)
- docs-generation: extract shared buildLinkRewriter() into docsUtils.ts; both
  electronDocs.ts and tauriDocs.ts now call the shared helper with their local
  allDocs and PUBLISHED_URL_PREFIX instead of duplicating the implementation
